[ refactor ] [!] useEmail 리팩토링
